PyMunk/requirements.txt
mediapipe>=0.10.0
numpy>=1.24.0
opencv-python>=4.8.0
pygame>=2.5.0
pymunk>=6.4.0
python-dotenv>=1.0.0
